Make the parser return a hash
I think this is what I want to evaluate an expression against.  Not sure
if I can use Parslet's Transform or not, since I don't see how to pass
in an array of auths to evaluate against.  May do the evaluation outside
of Parslet at this point.

Tests are broken now, need to figure out a better way to test.