This is really clean now
Parse will return a tree that I can now use to evaluate against an array
of auths to determine whether those auths are allowed or not allowed
access.

Need to fill out the README.  This is should not be used to protect
data, it is only a toy application to show how one could test auths in
data the user was returned.